Removed inclusion of chowning section
Removed the inclusion of the bootable/common/chowning.xml section from
all books.  It happens too early on embedded since we don't boot the
target until pretty much the end of the book.  If you perform the chowning
and then want to do anything else (fix something, do bclfs, etc), it has
to be done as root rather than the "clfs" user.  That's not a good idea,
too much risk of typing the wrong path and messing up the host.

This does not delete the bootable/common/chowning.xml file, just stops it
from being included.