Added ${CLFS}-final/build dir to be rm'ed
In the Copy to Target section, a $CLFS/build directory is created.
Should this not be the first time you do the Remove Unnecessary Files
section, a $CLFS/build directory will exist.  Need to ensure it gets
deleted from ${CLFS}-final/ since it contains tarballs that aren't needed
on the final system.