Revert "Install linux headers to cross-tools/include"
This reverts commit df099acd5ca16dcd82aef08badc46c5797eac408.

Moving the headers has uninteded side effects that aren't as easy to
overcome as originally thought.  Making gcc-final and uClibc properly
use headers located in ${CLFS}/cross-tools is not as straight forward as
hoped.  Plus, we delete all the headers before deploying to the target
anyway, so there's not much added size to the final system.